requirements.txt
affine==2.4.0
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.1.1
cycler==0.12.1
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-gis==1.0
dnspython==2.4.2
docopt==0.6.2
filelock==3.12.4
fiona==1.9.5
fonttools==4.42.1
GDAL==3.0.4
GeoAlchemy==0.7.2
geopandas==0.13.2
geoserver-rest==2.5.3
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.0
Js2Py==0.74
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==6.2.0
pipwin==0.5.2
postgres-helper==0.0.5
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2==2.9.9
pygments==2.17.2
PyGreSQL==6.0
pyjsparser==2.7.1
pymongo==4.6.1
pyparsing==3.1.1
PyPrind==2.11.3
pyproj==3.5.0
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2023.3.post1
rasterio==1.3.9
requests==2.31.0
Rtree==1.1.0
seaborn==0.12.2
shapely==2.0.2
six==1.16.0
snuggs==1.4.7
soupsieve==2.5
SQLAlchemy==2.0.24
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
xmltodict==0.13.0
zipp==3.17.0
